==> notebooks/0_1 Data Retrieval.ipynb <==
from pynytimes import NYTAPI

nyt_api = NYTAPI("SoAGIoyWawUiv8CcxLFD4MSXYckreA0d", parse_dates=True)

import datetime
import random
import pandas as pd
import time

def fetch_nyt_articles(start_date, end_date, file_name):
    try:
        articles = nyt_api.article_search(
            query="", 
            results=50,
            dates={
                "begin": start_date,
                "end": end_date
            },
            options={
                "sort": "oldest",
                "sources": [
                    "New York Times",
                    "AP",
                    "Reuters",
                    "International Herald Tribune"
                ],
                "document_type": [
            "article"
        ]
            }
        )

        df = pd.DataFrame(articles)
        df.to_csv(file_name, index=False)
        print(f"Articles saved to {file_name}")
        return df

    except Exception as e:
        print(f"An error occurred: {e}")
        return None
