==> notebooks/0_2 Data cleaning.ipynb <==
import pandas as pd
import numpy as np

df = pd.read_csv('combined_nyt_articles.csv')

df.sample(5)

df.shape

df['source'].value_counts()

df.info()

df.columns

df.duplicated().sum()

df.drop_duplicates(inplace=True)
df.duplicated().sum()

df.shape

df.isna().sum()

columns_to_drop = ['snippet', 'print_section', 'print_page', 'source', 'subsection_name', 'slideshow_credits']
df = df.drop(columns=columns_to_drop)

df.shape

df.isna().sum()

df = df.dropna()

df.shape

df.isna().sum()

df.head()

df.to_csv('cleaned_dataset_se.csv', index=False)

df2 = pd.read_csv('cleaned_dataset_se.csv')

df2.shape

df2["document_type"].value_counts() 

df2 = df2[df2["document_type"] == 'article']

df2.shape

df2["type_of_material"].value_counts()

df2.to_csv('cleaned_dataset_fo.csv', index=False)

df3 = pd.read_csv('cleaned_dataset_th.csv')

df4 = pd.read_csv('cleaned_dataset_fo.csv')

df3.shape

df4.shape

import pandas as pd
import glob

# Pattern to match the files. Adjust the pattern to match your files' naming convention
file_pattern = 'cleaned_dataset_*.csv'
csv_files = glob.glob(file_pattern)

# List to hold DataFrames
dfs = []

# Iterate over the files and append to the list
for file in csv_files:
    df = pd.read_csv(file)
    dfs.append(df)

# Concatenate all DataFrames into one
combined_df = pd.concat(dfs, ignore_index=True)

# Save the combined DataFrame to a new CSV file
combined_df.to_csv('combined_dataset.csv', index=False)

df_com = pd.read_csv('combined_dataset.csv')

df_com.shape

df_com["type_of_material"].value_counts()

#news_rows = df_com[df_com["type_of_material"] == "News"]
#reduced_news_rows = news_rows.iloc[1500:]  # Skip the first 1500
#other_rows = df_com[df_com["type_of_material"] != "News"]
#df_com_reduced = pd.concat([reduced_news_rows, other_rows])
