==> notebooks/0_3 Data Exploring.ipynb <==
import pandas as pd
import numpy as np

df = pd.read_csv('cleaned_dataset.csv')

df.shape

df.head(2)

df.info()

df["document_type"].value_counts()  # we may only keep the article 

## Supervised Learning: Text Classification

### Explore potential labels to be predicted

There are three features that can be used to as label: news_desk, section_name, type_of_material.

To keep the number of labels minimal, I have decided to use 'type_of_material' as the label feature. This feature will predict the material type of the article.

I have also chosen to retain only the articles with the 'document_type,' which has reduced the dataset size from 4103 to 3893. I am considering acquiring more data to expand the dataset further.

df["news_desk"].nunique() 

df["section_name"].nunique() 

df["type_of_material"].nunique() 

df["type_of_material"].value_counts() 

df_media = df[df["document_type"] == "multimedia"]

df_media["type_of_material"].value_counts()

df = df[df["document_type"] == 'article'] #keep only articles

df.shape

df["type_of_material"].value_counts() #the distribution of labels shows the imbalance of the dataset. 

The distribution of labels shows the dataset is quite imbalanced, with "News" being the dominant class. If we want to avoid the imbalanced data bias, I can always retrieve more data to increase the representation of the minority classes. 

Alternatively, there are several other methods that can be employed to address this issue of imbalanced classes. For example, combining all of the classes except "News" into a single category called "non_news" class. which we will be doing binary classification. 


### Explore the potential predictors


I print out the content of “abstract”, “lead_paragraph”, “headline”, “keywords” of the second row (the second article). 

I attempt to determine whether I just use one of the contents or the combination of all. 

To better understand if the contents of those can be used to predict the article's material type, I read the entire second article to see how closely these contents relate to the article's overall topic. 

print(df.loc[1, 'abstract'])

print(df.loc[1, 'lead_paragraph'])

print(df.loc[1, 'headline'])

print(df.loc[1, 'keywords'])

print(df.loc[1, 'web_url']) #to examine the entire article. 

Here is my findings after reading the contents of the four features and the full text of the article: 

This article's material type is marked as obituary.

Abstract: reflect the article's focus on personality, but may not always clearly indicate the material type of the article.

Lead paragraph: This feature is more descriptive, making it a strong candidate for identifying the material type.

Headline: Headlines are designed to give a quick summary of the article's content. In this case, it effectively indicates that the article is an obituary.

Keywords: Keywords are very indicative of the content and context. The presence of "Deaths (Obituaries)" is a strong predictor for the obituary category.

Overall context: after reading the entire article, confirming that these features are closely related to the overall topic and are good predictors of the material type. 

Conclusion: we will combine these features to make a robust predictor. or we can experiment with different combinations(like using all features or using two of the features) to see which yields the best classification results. 
Technically, we will combine these 4 features into a single feature as predictor. 




## Unsupervised Learning: Topic Modeling.

Goal: uncover hidden thematic structures in the data


For topic modeling, we will use the same textual features(the 4 features)as in text classification, but we won't need the label feature. As with text classification, we will also experiment with different combination of the 4 features to see which yields the most coherent and distinct topics. 

### Up to this point, we have selected our features for both supervised and unsupervised learning tasks.

## Feature engineering

Steps used to get from raw dataset to the final features used. which include data preprocessing, feature extraction (NLP like TF-IDF)


The preprocessing steps are similar for both tasks(text classificatin & topic modeling): 

lowercasing, removing special characters, tokenization, removing stop words, possibly stemming or lemmatization. 

## Project
name/title, goals, motivation, closest example

### Title
I am thinking of two titles: 

1, Dual Perspectives: Classifying and Theming NYTimes Articles.

2, The Article Analyst: Machine Learning Insights into NYTimes Journalism.




### Goals
Text Classification (Supervised Learning):
Develop a binary text classification model to categorize articles from The New York Times into two classes: "News" and "Non-News.

Topic Modeling (Unsupervised Learning):
Utilize topic modeling techniques to uncover and analyze the latent thematic structure in the dataset of articles from The New York Times.

General Project Objective

The overarching objective of this project is to leverage machine learning techniques to extract meaningful insights from a dataset of articles from The New York Times. This involves categorizing articles into relevant classes and discovering underlying themes, thereby enabling a deeper understanding of the content and structure of the news articles.

### Motivation

our motivation for undertaking this project is to harness the power of machine learning and natural language processing to deepen our understanding of news content. By classifying and analyzing New York Times articles, we aim to improve the accessibility and navigation of news media, uncovering underlying themes and patterns in journalistic content. This project represents a convergence of technological innovation and media analysis, offering valuable insights for both readers and industry professionals, and contributing to the evolving landscape of data-driven journalism.

### Closest Example

we are asked to find a closest example that is most similar to what we are proposing(stated in our proposal instruction)

The closest example is the course 655 NLP assignments: text classification,which using the English Wikipedia articles, a text of person's Wikipedia biography, determine the person's nationality. 

## Next Steps

Up to this point, I haven't delved into considering the supervised and unsupervised learning approaches, which encompass model selection, evaluation, and visualization methods. We can only contemplate these aspects after deciding on our specific supervised and unsupervised tasks.

The next immediate step I will take is to retrieve more data and prepare it in the format of our current dataset, just in case we need additional data in the future.
