==> notebooks/0_3 Data PreProcessing.ipynb <==
# INPUT = "../api/feb2024dataset/articles.csv"
# OUTPUT = "../api/feb2024dataset/processed_articles.csv"
INPUT = "../dataset/combined_dataset.csv"
OUTPUT = "processced df.csv"

import pandas as pd
import numpy as np
import ast

df = pd.read_csv(INPUT)

df.head(1)

df.columns

col_ke = ['abstract', 'lead_paragraph', 'headline', 'keywords', 'type_of_material', 'news_desk', 'section_name', 'byline']

df= df[col_ke]

df.head()

print(df.loc[1, 'headline'])

print(df.loc[1, 'keywords'])

### Extract value info from "headline" and "keywords"

#extract value texts from "headline" and "keywords"
def string_to_dict(string):
    try:
        return ast.literal_eval(string)
    except ValueError:
        return {}
    
def string_to_list_of_dicts(string):
    try:
        return ast.literal_eval(string)
    except ValueError:
        return []
#for the 'headline', extract the value associated with the 'main' key.
df['headline_text'] = df['headline'].apply(lambda x: string_to_dict(x).get('main', ''))

#for the "keywords", extract the "value" key
df['keywords_text'] = df['keywords'].apply(lambda x: ' '.join([kw.get('value', '') for kw in string_to_list_of_dicts(x)]))

df = df.drop(["headline", "keywords"], axis = 1)
df.head()

### Concatenation of Text Features

df['combined_text'] = df['abstract'].astype(str) + ' ' + df['lead_paragraph'].astype(str) + ' ' + df['headline_text'].astype(str) + ' ' + df['keywords_text'].astype(str)

df = df.drop(['abstract', 'lead_paragraph', 'headline_text', 'keywords_text'], axis=1)

df.tail()

print(df.loc[1, 'combined_text'])

df["news_desk"].value_counts()

df["news_desk"].unique()

df["section_name"].unique()

df["section_name"].value_counts()

### Consolidate Rare Categories in "news_desk


#Consolidate Rare Categories in "news_desk"

df['news_desk'] = df['news_desk'].replace({
    'Sports / Soccer': 'Sports',
    'Business / Energy & Environment': 'Business',
    'NODESK': 'Unknown',
    '0': 'Unknown',
    'Styles': 'Style & Fashion',
    'Fashion & Style': 'Style & Fashion',
    'TStyle': 'Style & Fashion',
    'Arts&Leisure': 'Arts & Culture',
    'Culture': 'Arts & Culture',
    'TCulture': 'Arts & Culture',
    'Editorial': 'Opinion & Editorial',
    'OpEd': 'Opinion & Editorial',
    'Opinion': 'Opinion & Editorial',
    'Dining': 'Lifestyle & Leisure',
    'Travel': 'Lifestyle & Leisure',
    'Well': 'Lifestyle & Leisure',
    'Home': 'Lifestyle & Leisure'
})

# Handling infrequent categories
infrequent_desks = df['news_desk'].value_counts()[df['news_desk'].value_counts() < 5].index
df['news_desk'] = df['news_desk'].replace(infrequent_desks, 'Other')


df["news_desk"].value_counts()

print(df.loc[1, 'byline'])

### Extract Author's name from "byline"



def extract_author(byline):
    if pd.isna(byline) or not byline:
        return ''  # Return empty string if byline is NaN or empty

    try:
        byline_dict = ast.literal_eval(byline)
        if 'person' in byline_dict and byline_dict['person']:
            person_info = byline_dict['person'][0]
            first_name = person_info.get('firstname', '')
            last_name = person_info.get('lastname', '')
            return f"{first_name} {last_name}".strip()
        elif 'original' in byline_dict and byline_dict['original']:
            return byline_dict.get('original', '').split('By ')[-1].strip()
        else:
            return ''
    except (ValueError, SyntaxError):
        return ''

df['author'] = df['byline'].apply(extract_author)

df = df.drop(columns=['byline'])

df.head()

df["author"].value_counts()

df.to_csv(OUTPUT, index=False)
